+++ web/src/main/java/com/homeapp/javatraining/exception/user/InvalidCredentialsException.java
@@ -1,0 +1,10 @@
+package com.homeapp.javatraining.exception.user;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class InvalidCredentialsException extends ApiException {
+
+    public InvalidCredentialsException() {
+        super("Invalid username or password");
+    }
+}